+++ src/components/Modal.tsx
@@ -1,12 +1,11 @@
 import ReactModal from "react-modal";
-import { useAppDispatch, useAppSelector } from "../redux/hooks";
-import { deleteItem } from "../redux/taskSlice";
-import { toggleShowModal } from "../redux/modalSlice";
+import { useAppSelector } from "../redux/hooks";
+import { DeleteButton } from "./DeleteButton";
+import { CancelDeleteButton } from "./CancelDeleteButton";
 import "../css/modal.css";
 
 export function Modal() {
   const modalIsOpen = useAppSelector((state) => state.modal.modalIsOpen);
-  const dispatch = useAppDispatch();
   const customModalStyles = {
     content: {
       inset: "0",
@@ -16,15 +15,6 @@ export function Modal() {
     },
   };
 
-  function deleteHandler() {
-    dispatch(deleteItem());
-    dispatch(toggleShowModal());
-  }
-
-  function closeModal() {
-    dispatch(toggleShowModal());
-  }
-
   return (
     <ReactModal
       isOpen={modalIsOpen}
@@ -32,17 +22,9 @@ export function Modal() {
       ariaHideApp={false}
     >
       <h2>Are you sure delete?</h2>
-      <div className="btnBox">
-        <button
-          className="btn btn-delete"
-          type="button"
-          onClick={deleteHandler}
-        >
-          <span>Delete</span>
-        </button>
-        <button className="btn btn-cancel" type="reset" onClick={closeModal}>
-          <span>Cancel</span>
-        </button>
+      <div className="btnContainer">
+        <DeleteButton />
+        <CancelDeleteButton />
       </div>
     </ReactModal>
   );
